fix: Override Git config core.abbrev when reading version from SCM
The configuration represents the length commit hashes are abbreviated
to. As pdm-backend parses output of git describe to obtain the package
version, this configuration affects format of package versions that must
be represented with a commit hash part, which is quite surprising and
may introduce reproducible issues.

Let's override core.abbrev with auto and let git decide how long the
hash should be, keeping the format of generated version consistent
regardless of value of core.abbrev and fixing failures of
test__get_version_from_scm__returns_default_if_tag_cannot_be_parsed[git]
that occur when the git configuration is set globally.

Closes: https://github.com/pdm-project/pdm/issues/3490
Signed-off-by: Yao Zi <ziyao@disroot.org>